+++ ft_substr.c
@@ -8,9 +8,9 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	i = 0;
 	if (!s)
 		return (0);
-	if ((size_t)ft_strlen(s) < start)
+	if (ft_strlen(s) < start)
 		return (ft_strdup(""));
-	if ((size_t)ft_strlen(s) < len)
+	if (ft_strlen(s) < len)
 		substr = (char *)malloc((ft_strlen(s) - start + 1) * sizeof(char));
 	else
 		substr = (char *)malloc((len + 1) * sizeof(char));
